Extract media type to its own file
