+++ packages/alfa-media/src/media.ts
@@ -20,6 +20,7 @@ import { Slice } from "@siteimprove/alfa-slice";
 import * as json from "@siteimprove/alfa-json";
 
 import { Resolver } from "./resolver";
+import { Type } from "./type";
 
 const {
   map,
@@ -66,70 +67,6 @@ export namespace Media {
     LessThan = "<",
     LessThanEqual = "<=",
   }
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#media-type
-   */
-  export class Type implements Queryable<Type.JSON> {
-    public static of(name: string): Type {
-      return new Type(name);
-    }
-
-    private readonly _name: string;
-
-    private constructor(name: string) {
-      this._name = name;
-    }
-
-    public get name(): string {
-      return this._name;
-    }
-
-    public matches(device: Device): boolean {
-      switch (this._name) {
-        case "screen":
-          return device.type === Device.Type.Screen;
-
-        case "print":
-          return device.type === Device.Type.Print;
-
-        case "speech":
-          return device.type === Device.Type.Speech;
-
-        case "all":
-          return true;
-
-        default:
-          return false;
-      }
-    }
-
-    public equals(value: unknown): value is this {
-      return value instanceof Type && value._name === this._name;
-    }
-
-    public toJSON(): Type.JSON {
-      return {
-        name: this._name,
-      };
-    }
-
-    public toString(): string {
-      return this._name;
-    }
-  }
-
-  export namespace Type {
-    export interface JSON {
-      [key: string]: json.JSON;
-      name: string;
-    }
-  }
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-type
-   */
-  const parseType = map(Token.parseIdent(), (ident) => Type.of(ident.value));
 
   /**
    * @see https://drafts.csswg.org/mediaqueries/#media-feature
@@ -999,7 +936,7 @@ export namespace Media {
       pair(
         pair(
           option(delimited(option(Token.parseWhitespace), parseModifier)),
-          parseType
+          Type.parse
         ),
         option(
           right(
